Commit for Issue #969
Changed Auto Layout Constraints
