+++ plugins/de.ovgu.featureide.fm.ui/src/de/ovgu/featureide/fm/ui/editors/featuremodel/layouts/ManualLayout.java
@@ -20,9 +20,12 @@
  */
 package de.ovgu.featureide.fm.ui.editors.featuremodel.layouts;
 
+import de.ovgu.featureide.fm.core.base.FeatureUtils;
+import de.ovgu.featureide.fm.ui.editors.FeatureUIHelper;
 import de.ovgu.featureide.fm.ui.editors.IGraphicalConstraint;
 import de.ovgu.featureide.fm.ui.editors.IGraphicalFeature;
 import de.ovgu.featureide.fm.ui.editors.IGraphicalFeatureModel;
+import de.ovgu.featureide.fm.ui.properties.FMPropertyManager;
 
 /**
  * Layouts the features at the feature diagram using their saved Positions.
@@ -42,8 +45,23 @@ public class ManualLayout extends FeatureDiagramLayoutManager {
 		for (final IGraphicalFeature feature : featureModel.getVisibleFeatures()) {
 			setLocation(feature, feature.getLocation());
 		}
-		for (final IGraphicalConstraint constraint : featureModel.getVisibleConstraints()) {
-			constraint.setLocation(constraint.getLocation());
+		if (featureModel.getLayout().isAutoLayoutConstraints()) {
+			final IGraphicalFeature root =
+				FeatureUIHelper.getGraphicalFeature(FeatureUtils.getRoot(featureModel.getFeatureModelManager().getSnapshot()), featureModel);
+			// Calculate yoffset as the position of the lowest visible feature + its height.
+			int yoffset = FMPropertyManager.getLayoutMarginY();
+
+			for (final IGraphicalFeature feat : featureModel.getVisibleFeatures()) {
+				if (yoffset < feat.getLocation().y) {
+					yoffset += FMPropertyManager.getFeatureSpaceY();
+				}
+			}
+
+			layoutConstraints(yoffset, featureModel.getVisibleConstraints(), getBounds(root));
+		} else {
+			for (final IGraphicalConstraint constraint : featureModel.getVisibleConstraints()) {
+				constraint.setLocation(constraint.getLocation());
+			}
 		}
 	}
